Bug Fixes: - Fix pattern detection threshold (> 0.5 -> > 0) that caused 3-panel   visualization to miss most primes. When grids are normalized, single   primes get low values (e.g., 0.33) that the > 0.5 threshold missed.   Fixed in run_autonomous_discovery.py (5 locations) and   enhanced_detection.py (12 locations).
- Fix logarithmic spiral overflow warning by clamping exponent to 700
  before calling np.exp() in logarithmic_coords function.

- Fix 60+ mypy type errors across the codebase:
  - Optional parameters with None defaults now use Optional[] wrapper
  - Renamed variables to avoid type conflicts when reassigning classes
  - Added Union types for polymorphic module attributes in models.py
  - Separated calculations from Optional typing in predictor.py
  - Added missing List/Dict/Tuple type annotations

Documentation:
  - README.md: Added Recent Updates section
  - doc/ARCHITECTURE.md: Added Critical Bug Fixes section
  - doc/OUTPUT_ANALYSIS.md: Added Section 17 documenting fixes

+++ src/prime_plot/evaluation/detectors.py
@@ -222,14 +222,14 @@ def detect_clusters(
     # Score based on cluster quality
     if clusters:
         # Good clustering = many well-defined clusters with consistent density
-        avg_density = np.mean([c.density for c in clusters])
-        density_consistency = 1 - np.std([c.density for c in clusters]) / (avg_density + 1e-6)
-        density_consistency = max(0, density_consistency)
+        avg_density = float(np.mean([c.density for c in clusters]))
+        density_consistency = 1 - float(np.std([c.density for c in clusters])) / (avg_density + 1e-6)
+        density_consistency = max(0.0, density_consistency)
 
         # Coverage = fraction of points in clusters
         total_points = binary.sum()
         clustered_points = sum(c.size for c in clusters)
-        coverage = clustered_points / total_points if total_points > 0 else 0
+        coverage = clustered_points / total_points if total_points > 0 else 0.0
 
         # Number of clusters (more = more structured, up to a point)
         num_score = min(len(clusters) / 20, 1.0)
@@ -239,7 +239,7 @@ def detect_clusters(
     else:
         score = 0.0
 
-    return clusters, score
+    return clusters, float(score)
 
 
 def compute_fft_spectrum(
